# docs/swagger.yaml
definitions:
  models.AssignRoleRequest:
    properties:
      role:
        type: string
    type: object
  models.BadRequestResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Bad request. Missing or invalid required fields.
        type: string
    type: object
  models.ConflictResponse:
    properties:
      code:
        example: 409
        type: integer
      message:
        example: Conflict. Resource already exists or operation conflicts with current
          state.
        type: string
    type: object
...
  models.InternalServerErrorResponse:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: Internal server error occurred.
        type: string
    type: object
...
  models.ResetPasswordRequest:
    properties:
      email:
        type: string
    type: object
  models.ServiceResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Operation completed successfully
        type: string
    type: object
  models.StatusUnprocessableEntityResponse:
    properties:
      code:
        example: 422
        type: integer
      message:
        example: Invalid request body format. Please check your JSON syntax and field
          types.
        type: string
    type: object
  models.UnauthorizedResponse:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized access. Please provide valid authentication credentials.
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  models.UserRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.UserVerifyOTPRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      new_password:
        type: string
      otp:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user from the system using their unique identifier
      parameters:
      - description: User ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User successfully deleted
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Delete a user
      tags:
      - User
    get:
      description: Fetches the details of a user by their unique user ID. This endpoint
        is typically used by admins or services that need to look up users.
      parameters:
      - description: Unique identifier of the user to retrieve
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details successfully retrieved
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get User Details by ID
      tags:
      - User
  /user/{id}/role:
    post:
      consumes:
      - application/json
      description: Assigns a specific role to a user identified by their user ID.
        Only users with sufficient privileges (e.g., admins) can perform this action.
      parameters:
      - description: Unique identifier of the user to assign a role
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role assigned successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "422":
          description: Unprocessable entity, invalid input
          schema:
            $ref: '#/definitions/models.StatusUnprocessableEntityResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Assign Role to User
      tags:
      - User
  /user/details:
    get:
      description: Returns the details of the user currently authenticated via the
        API key or token. This endpoint is useful for profile pages or user dashboards.
      produces:
      - application/json
      responses:
        "200":
          description: User details successfully retrieved
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Authenticated User Details
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Allows the authenticated user to update their profile information
        such as name, email, or other editable fields. Requires authentication.
      parameters:
      - description: Fields to update for the user profile
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User details updated successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "422":
          description: Unprocessable entity, invalid input
          schema:
            $ref: '#/definitions/models.StatusUnprocessableEntityResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Authenticated User Details
      tags:
      - User
  /user/password/reset:
    post:
      consumes:
      - application/json
      description: Initiates the password reset process for a user by sending a reset
        link or OTP to the user's email.
      parameters:
      - description: Password reset request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset initiated successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "422":
          description: Unprocessable entity, invalid input
          schema:
            $ref: '#/definitions/models.StatusUnprocessableEntityResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Reset User Password
      tags:
      - User
  /user/password/set:
    post:
      consumes:
      - application/json
      description: Sets a new password for the user after verifying the OTP sent to
        their email. Requires email, OTP, new password, and confirmation.
      parameters:
      - description: OTP verification and new password details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserVerifyOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "409":
          description: Conflict, password confirmation failed
          schema:
            $ref: '#/definitions/models.ConflictResponse'
        "422":
          description: Unprocessable entity, invalid input
          schema:
            $ref: '#/definitions/models.StatusUnprocessableEntityResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Set New User Password
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: Registers a new user in the system under a specific tenant. This
        endpoint is typically used for onboarding new users. Requires all mandatory
        fields such as name, email, and password.
      parameters:
      - description: User registration details including name, email, and password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "409":
          description: Conflict, user already exists
          schema:
            $ref: '#/definitions/models.ConflictResponse'
        "422":
          description: Unprocessable entity, invalid input
          schema:
            $ref: '#/definitions/models.StatusUnprocessableEntityResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      summary: Register New User
      tags:
      - User
  /users:
    get:
      description: Fetches a paginated list of all users in the system. This endpoint
        is typically used by admins to view and manage users. Returns user details
        with pagination support.
      produces:
      - application/json
      responses:
        "200":
          description: Users list successfully retrieved with pagination
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List All Users
      tags:
      - User
  /users/{id}/disable:
    put:
      description: Disables a user account, preventing them from accessing the system.
        Only users with sufficient privileges can perform this action.
      parameters:
      - description: Unique identifier of the user to disable
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User account disabled successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disable User Account
      tags:
      - User
  /users/{id}/enable:
    put:
      description: Enables a previously disabled user account, allowing them to access
        the system again. Only users with sufficient privileges can perform this action.
      parameters:
      - description: Unique identifier of the user to enable
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User account enabled successfully
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "400":
          description: Bad request, missing required fields
          schema:
            $ref: '#/definitions/models.BadRequestResponse'
        "401":
          description: Unauthorized, invalid or missing authentication
          schema:
            $ref: '#/definitions/models.UnauthorizedResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ServiceResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.InternalServerErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Enable User Account
      tags:
      - User
swagger: "2.0"